scraper/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
playwright==1.40.0
supabase==2.0.2
python-dotenv==1.0.0
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.1
pandas==2.1.3
numpy==1.25.2
Pillow==10.1.0
PyPDF2==3.0.1
openpyxl==3.1.2
